Switch to tsc
- Replace `babel` with `tsc`. Afterall we use TypeScript, why not use all its features.
- Generate source maps
- Move exported types to `types.d.ts`
- Amend `mocha` configuration to use `ts-node`.
- Amend `nyc` configuration. Follow [nyc-config-typescript instructions](https://www.npmjs.com/package/@istanbuljs/nyc-config-typescript)
- Add @nikolay-borzov as a maintainer in `package.json`
- Specify minimal supported node version in `package.json`
