--- .babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    "@babel/typescript",
-    "@babel/preset-env"
-  ],
-  "env": {
-    "test": {
-      "plugins": [
-        "istanbul"
-      ]
-    }
-  }
-}
