+++ .mocharc.base.js
@@ -1,6 +1,7 @@
 module.exports = {
-  file: ['tests/setup.js'],
-  require: 'tests/babel-register.js',
+  file: 'tests/setup.ts',
+  require: ['ts-node/register', 'source-map-support/register'],
   timeout: 60000,
   colors: true,
+  recursive: true,
 };
