+++ .nycrc.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@istanbuljs/nyc-config-typescript",
+  "include": "src",
+  "exclude": [
+    "src/cli.ts",
+    "src/index.ts"
+  ]
+}
