+++ tests/unit/explore.test.ts
@@ -1,7 +1,8 @@
 import { expect } from 'chai';
 
 import { adjustSourcePaths } from '../../src/explore';
-import { FileDataMap, ExploreOptions } from '../../src';
+
+import { FileDataMap, ExploreOptions } from '../../src/types';
 
 interface AdjustSourcePathsTest {
   name: string;
