+++ tests/unit/html.test.ts
@@ -1,7 +1,8 @@
 import snapshot from '@smpx/snap-shot-it';
 
 import { getWebTreeMapData } from '../../src/html';
-import { FileDataMap } from '../../src/index';
+
+import { FileDataMap } from '../../src/types';
 
 describe('html', () => {
   describe('getWebTreeMapData', () => {
